Port the test infrastructure to Tasty
(This is WIP) [ci skip]

`runtest.hs` is replaced by `TestRun.hs`.
This program imports `TestData.hs` which holds the tests metadata,
then turns it into tests usable by Tasty.

The functional changes are the following:
- The tasty framework provides a pretty output by default
  (colors, individual timings, proper indentation).
  Tests can run in parallel but the output is still deterministic.
- Tasty.Golden does the heavy lifting,
  i.e. comparing outputs and measuring time spent.
- Output is captured through a pipe connected to a bash process running
  the `run` script for each test ; this is implemented by the `runTest`
  function
- Tests are organised in families, reflecting directories' prefixes.
  They are listed in `testFamiliesData` in `TestData.hs`.
- Each test must be registered in `testFamiliesData` under its own
  family.
- A custom option has been implemented to run the test suite with the node
  code generator. The list of compatible code generators for each test
  is provided in `testFamiliesData`.
- Tasty.Rerun provides rudimentary incremental testing.

The previous Makefile in `test` has disappeared in favour of passing
arguments to the test program.

Note that this commit disables testing because a compatibility
layer for `cabal test` must be implemented first. This is necessary
because the test program requires external dependencies.
However, you should be able to run the tests with `stack test`.